Update Result Post

+++ app/src/main/java/com/b21_cap0183/paddycare/core/data/source/local/entity/ResultEntity.kt
@@ -10,7 +10,7 @@ data class ResultEntity(
     @PrimaryKey
     @NonNull
     @ColumnInfo(name = "resultId")
-    var resultId: Int,
+    var resultId: Long,
 
     @ColumnInfo(name = "resultName")
     var resultName: String,
