+++ app/src/main/java/com/b21_cap0183/paddycare/core/domain/model/Result.kt
@@ -5,7 +5,7 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class Result(
-    var resultId: Int,
+    var resultId: Long,
     var resultName: String,
     var resultDesc: String,
     var resultSolution: String
